% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyscrape.R
\name{scrape}
\alias{scrape}
\title{scrape urls with proxy (Tor as default)}
\usage{
scrape(
  urls,
  proxy = "socks5://127.0.0.1",
  port = 9050,
  max_retry = 3,
  useragent = NULL,
  random_ua = TRUE
)
}
\arguments{
\item{urls, }{character vector of urls}

\item{proxy, }{character, use default Tor proxy}

\item{port, }{int (default to 9050), use default Tor port 9050}

\item{max_retry, }{int (default 5), how many times to retry on failure}

\item{useragent, }{character (default NULL), overwrite useragent if supplied}

\item{random_ua, }{bool (default TRUE), random user agent by Randomuseragent::random_useragent()}
}
\description{
scrape urls with proxy (Tor as default)
}
